Bluetooth: classic: Move classic source to sub folder
Move source code of classic from bluetooth.h to
./classic/classic.h.
